Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0383-ransom-note/0383-ransom-note.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool canConstruct(string ransomNote, string magazine) {
         
-        int alpha[26];
+        short alpha[26];
 
         for(char c : ransomNote)
         {
